// frontend/src/app/api/model/recipeDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RecipeDto { 
    id?: string;
    name: string;
    description?: string;
    cookingTimeMinutes?: number;
    instructions: string;
    dishId: string;
}